requirements.txt
# aiohappyeyeballs==2.6.1
# aiohttp==3.12.14
# aiosignal==1.4.0
# annotated-types==0.7.0
# anyio==4.9.0
# attrs==25.3.0
# certifi==2025.7.14
# charset-normalizer==3.4.2
# click==8.2.1
# dataclasses-json==0.6.7
# distro==1.9.0
# fastapi==0.116.1
# frozenlist==1.7.0
# groq==0.30.0
# h11==0.16.0
# httpcore==1.0.9
# httpx==0.28.1
# httpx-sse==0.4.1
# idna==3.10
# jsonpatch==1.33
# jsonpointer==3.0.0
# langchain==0.3.27
# langchain-community==0.3.27
# langchain-core==0.3.72
# langchain-groq==0.3.6
# langchain-tavily==0.2.11
# langchain-text-splitters==0.3.9
# langgraph==0.5.4
# langgraph-checkpoint==2.1.1
# langgraph-prebuilt==0.5.2
# langgraph-sdk==0.1.74
# langsmith==0.4.8
# marshmallow==3.26.1
# multidict==6.6.3
# mypy_extensions==1.1.0
# numpy==2.3.2
# orjson==3.11.1
# ormsgpack==1.10.0
# packaging==25.0
# propcache==0.3.2
# pvporcupine==3.0.5
# PyAudio==0.2.14
# pydantic==2.11.7
# pydantic-settings==2.10.1
# pydantic_core==2.33.2
# PyMuPDF==1.26.3
# python-dotenv==1.1.1
# python-multipart==0.0.20
# PyYAML==6.0.2
# requests==2.32.4
# requests-toolbelt==1.0.0
# sniffio==1.3.1
# SpeechRecognition==3.14.3
# SQLAlchemy==2.0.41
# starlette==0.47.2
# tenacity==9.1.2
# typing-inspect==0.9.0
# typing-inspection==0.4.1
# typing_extensions==4.14.1
# urllib3==2.5.0
# uvicorn==0.35.0
# xxhash==3.5.0
# yarl==1.20.1
# zstandard==0.23.0



fastapi
uvicorn
python-multipart
PyMuPDF
langgraph
langchain-core
langchain-groq
python-dotenv
typing-extensions

